+++ src/main/java/org/geowe/client/local/layermanager/tool/create/GeoDataImportDialog.java
@@ -41,6 +41,7 @@ import com.google.gwt.event.dom.client.ChangeHandler;
 import com.google.gwt.event.dom.client.KeyCodes;
 import com.google.gwt.event.logical.shared.SelectionEvent;
 import com.google.gwt.event.logical.shared.SelectionHandler;
+import com.google.gwt.http.client.URL;
 import com.google.gwt.i18n.client.HasDirection.Direction;
 import com.google.gwt.user.client.ui.Anchor;
 import com.google.gwt.user.client.ui.HasHorizontalAlignment;
@@ -300,13 +301,15 @@ public class GeoDataImportDialog extends Dialog {
 				urlShared.setVisible(true);
 			}
 
-			private String getHref() {
+		private String getHref() {
 				String baseUrl = GWT.getHostPageBaseURL();
-
-				return baseUrl + "?layerUrl=" + urlTextField.getValue()
+				
+				baseUrl +=  "?layerUrl=" +  URL.encodeQueryString(urlTextField.getValue())
 						+ "&layerName=" + getLayerName() + "&layerProj="
 						+ getProjectionName() + "&layerFormat="
 						+ getDataFormat();
+				
+				return baseUrl;
 			}
 		};
 	}