+++ src/main/webapp/js/EPSG.js
@@ -4,8 +4,14 @@ Proj4js.defs["EPSG:25830"] = "+proj=utm +zone=30 +ellps=GRS80 +units=m +no_defs"
 //Canary Islands
 Proj4js.defs["EPSG:32626"] = "+proj=utm +zone=26 +ellps=WGS84 +datum=WGS84 +units=m +no_defs";
 Proj4js.defs["EPSG:32627"] = "+proj=utm +zone=27 +ellps=WGS84 +datum=WGS84 +units=m +no_defs";
-//South Africa
+//Africa
+Proj4js.defs["EPSG:22032"] = "+proj=utm +zone=32 +south +ellps=clrk80 +units=m +no_defs";
+Proj4js.defs["EPSG:22033"] = "+proj=utm +zone=33 +south +ellps=clrk80 +units=m +no_defs ";
+Proj4js.defs["EPSG:32732"] = "+proj=utm +zone=32 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs";
 Proj4js.defs["EPSG:32733"] = "+proj=utm +zone=33 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs";
 Proj4js.defs["EPSG:32734"] = "+proj=utm +zone=34 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs";
 Proj4js.defs["EPSG:32735"] = "+proj=utm +zone=35 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs";
 Proj4js.defs["EPSG:32736"] = "+proj=utm +zone=36 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs";
+Proj4js.defs["EPSG:32737"] = "+proj=utm +zone=37 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs";
+Proj4js.defs["EPSG:3036"] = "+proj=utm +zone=36 +south +ellps=WGS84 +towgs84=0,0,0,-0,-0,-0,0 +units=m +no_defs";
+Proj4js.defs["EPSG:3037"] = "+proj=utm +zone=37 +south +ellps=WGS84 +towgs84=0,0,0,-0,-0,-0,0 +units=m +no_defs";